moved inv-diag runner, edit imports on setup
