+++ runzi/cli/inversion_diagnostic_runner.py
@@ -57,9 +57,9 @@ def inversion_diagnostic_runner(general_task_id):
 
     elif CLUSTER_MODE == EnvMode['AWS']:
         for script_or_config in scripts:
-            print('AWS_TIME!: ', script_or_config)
-            # res = batch_client.submit_job(**script_or_config)
-            # print(res)
+            # print('AWS_TIME!: ', script_or_config)
+            res = batch_client.submit_job(**script_or_config)
+            print(res)
 
     elif CLUSTER_MODE == EnvMode['CLUSTER']:
         for script_or_config in scripts:
